% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Lower_Seine_Rouen}
\alias{Lower_Seine_Rouen}
\title{Stage record and ADCP discharge measurements
of the Lower Seine river at Oissel and Petite-Couronne}
\format{
\describe{
\item{date}{Date with time}
\item{h1}{Real value, stage record upstream, at Oissel}
\item{u.h1}{Uncertainty of the stage upstream expressed as standard deviation}
\item{h2}{real value, stage record downstram, at Petite-Couronne}
\item{u.h2}{Uncertainty of the stage downstream expressed as standard deviation}
\item{Q}{Discharge ADCP measurement}
\item{u_Q}{Uncertainty of the discharge expressed as standard deviation}
}
}
\usage{
Lower_Seine_Rouen
}
\description{
A data frame containing information about the stage record in meters,
referenced to the Marine elevation of Le Havre from 28/09/2015 until 01/10/2015
provided by Stéphane Piney (SPC DREAL Normandy).
Additionally, information on ADCP discharge measurements is available
from the gauging campaign conducted on 29/09/2015S
}
\keyword{datasets}
