% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_Q_sim_Qmec}
\alias{plot_Q_sim_Qmec}
\title{Plot of the discharge simulation}
\usage{
plot_Q_sim_Qmec(Q_observed, Q_simulated)
}
\arguments{
\item{Q_observed}{data frame, discharge measurements}

\item{Q_simulated}{list or data frame, discharge simulation with their uncertainties}
}
\value{
plot_Q: ggplot, discharge simulation the associated uncertainties
}
\description{
Plot of the discharge simulation
}
\details{
Function need results from \code{Prediction_Q_Qmec} to plot of the discharge simulation
}
