% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{Prediction_Q_Qmec}
\alias{Prediction_Q_Qmec}
\title{Discharge simulation with uncertainties}
\usage{
Prediction_Q_Qmec(
  CalData.object,
  temp.folder = file.path(tempdir(), "BaM"),
  DoParam_Unc = TRUE,
  DoTotal_Unc = TRUE
)
}
\arguments{
\item{CalData.object}{object, calibration object obtained from \code{RBaM} package}

\item{temp.folder}{directory, temporary directory to write computations}

\item{DoParam_Unc}{logical, if \code{TRUE} propagate parametric uncertainty}

\item{DoTotal_Unc}{logical, if \code{TRUE} propagate total uncertainty}
}
\value{
ggplots, discharge simulation with uncertainties
}
\description{
Discharge simulation with uncertainties
}
